Move pytest fixtures into a reusable location
This moves and registers the pytest fixtures for this project into
a resuable location. These fixtures can now be used by external projects.
