+++ cnxdb/contrib/pytest.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import
 import psycopg2
 import pytest
